refactor: extract models and middleware from api/server.py (Phase 1+2)
- Extract 21 Pydantic model classes to api/models.py (259 lines)
- Extract 4 middleware classes to api/middleware.py (272 lines)
- server.py: 4,154 → 3,755 lines (-399, 9.6% reduction)
- Backward compat preserved: from agent_bom.api.server import app
- All 163 API tests pass, zero modifications to test files

Phase 1+2 of #584
